% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{ww_animate_hydrograph}
\alias{ww_animate_hydrograph}
\title{Hydrograph Animation}
\usage{
ww_animate_hydrograph(daily_values, dir = NULL)
}
\arguments{
\item{daily_values}{A previously create \code{whitewater::ww_dvUSGS()} object.}

\item{dir}{character. A path to store the GIF files. \code{tempfile()} (if NULL) or ex. \verb{some/file/path/}.}
}
\value{
A GIF.
}
\description{
Hydrograph Animation
}
\examples{
\dontrun{

library(whitewater.extras)
library(whitewater)
#provide a USGS site number that you want to look at visually (we need the data first)
kootenai_river <- ww_dvUSGS(sites = '12305000')

ww_animate_hydrograph(daily_values = kootenai_river)

}
}
