% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_addons.R
\name{ww_get_rolling}
\alias{ww_get_rolling}
\title{Rolling Stats}
\usage{
ww_get_rolling(data, vars, ...)
}
\arguments{
\item{data}{A data.frame.}

\item{vars}{A vector of columns names to mutate.}

\item{...}{values to pass to \code{slider::slide()}, e.g. \code{.after = 7}.}
}
\value{
Returns numerous stats (mean, median, sd) based on a rolling window.
}
\description{
Adds rolling stats based on window size.
}
\note{
The user must specify the \code{vars} used but has more flexibility with inputs to \code{slider::slide()}. In addition, this is assuming that
all of your data is complete and has no magic to know otherwise.
}
\examples{
rolling_flow <- data.frame(flow = c(seq(30, 60), seq(60, 30, length.out = 60)))
rolling_flow \%>\% ww_get_rolling(vars = 'Flow', .after = 7)
}
