% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ww_hysteresis_plot}
\alias{ww_hysteresis_plot}
\title{Hysteresis Plot}
\usage{
ww_hysteresis_plot(daily_values, value_name, lag_days = 1)
}
\arguments{
\item{daily_values}{A previously create \code{whitewater::ww_dvUSGS()} object.}

\item{value_name}{The variable to include as the y-axis.}

\item{lag_days}{Number of days to lag.}
}
\value{
A ggplot plot.
}
\description{
Hysteresis Plot
}
\examples{
\dontrun{

library(whitewater.extras)
library(whitewater)
#provide a USGS site number that you want to look at visually (we need the data first)
kootenai_river <- ww_dvUSGS(sites = '12305000')

ww_hysteresis_plot(daily_values = kootenai_river)

}
}
