% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ww_monthly_trends}
\alias{ww_monthly_trends}
\title{Monthly Trends}
\usage{
ww_monthly_trends(daily_values, value_name)
}
\arguments{
\item{daily_values}{A previously create \code{whitewater::wym_dvUSGS()} object.}

\item{value_name}{The variable to include as the y-axis.}
}
\description{
Monthly Trends
}
