% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ww_raster_hydrograph}
\alias{ww_raster_hydrograph}
\title{Raster Hydrograph}
\usage{
ww_raster_hydrograph(daily_values, value_name, without_ggfx = FALSE)
}
\arguments{
\item{daily_values}{A previously create \code{whitewater::ww_dvUSGS()} object.}

\item{value_name}{The variable to include as the y-axis.}

\item{without_ggfx}{logical. FALSE (default).}
}
\value{
A ggplot.
}
\description{
Raster Hydrograph
}
\examples{
\dontrun{

library(whitewater.extras)
library(whitewater)
#provide a USGS site number that you want to look at visually (we need the data first)
kootenai_river <- ww_dvUSGS(sites = '12305000')

ww_raster_hydrograph(daily_values = kootenai_river)

}
}
